CloudLab-Manager/backend/requirements.txt
flask
pyjwt
docker
